+++ honei/backends/cuda/eq_dist_grid.cu
@@ -31,7 +31,7 @@ namespace honei
                 float g, float e,
                 unsigned long offset, unsigned long size)
         {
-            int idx = (blockDim.y * blockIdx.y * gridDim.x * blockDim.x) + (blockDim.x * blockIdx.x) + threadIdx.x;
+            unsigned long idx = (blockDim.y * blockIdx.y * gridDim.x * blockDim.x) + (blockDim.x * blockIdx.x) + threadIdx.x;
             if (idx < size)
             {
                 unsigned long index(idx + offset);
